zeppelin/src/apps/core/api_views.py
from .models import (
    Historical,
    Base,
)
from .serializers import (
    HistoricalReadSerializer, HistoricalWriteSerializer,
    BaseReadSerializer, BaseWriteSerializer,
)

from rest_framework.viewsets import ModelViewSet
from rest_framework.permissions import IsAdminUser
from rest_condition import And, Or
from oauth2_provider.contrib.rest_framework import TokenHasReadWriteScope, OAuth2Authentication
from rest_framework.authentication import SessionAuthentication
from .pagination import CustomPagination
from rest_framework import generics
from rest_framework import filters
import django_filters.rest_framework


